feat(cli): add @teamclu/cli with marketplace publish and domain stubs
Single npm/npx entry (teamclu) for ops: marketplace publish is ready;
mcp / env / knowledge domains are reserved. Root marketplace:publish
forwards to the new package.
