+++ packages/cli/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@teamclu/cli",
+  "version": "0.1.0",
+  "description": "TeamClu ops CLI — marketplace publish today; mcp / env / knowledge later",
+  "type": "module",
+  "bin": {
+    "teamclu": "./bin/teamclu.mjs"
+  },
+  "files": [
+    "bin",
+    "src",
+    "README.md"
+  ],
+  "engines": {
+    "node": ">=20"
+  },
+  "license": "UNLICENSED",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/different-ai-studio/teamclu.git",
+    "directory": "packages/cli"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "keywords": [
+    "teamclu",
+    "marketplace",
+    "skills",
+    "cli"
+  ]
+}
