+++ packages/cli/src/_shared/api.mjs
@@ -1,0 +1,23 @@
+/** Tiny JSON admin client (x-webhook-secret). */
+
+export async function adminFetch(base, secret, method, path, body) {
+  const res = await fetch(`${base}${path}`, {
+    method,
+    headers: {
+      "content-type": "application/json",
+      "x-webhook-secret": secret,
+    },
+    body: body ? JSON.stringify(body) : undefined,
+  });
+  const text = await res.text();
+  let json = null;
+  try {
+    json = text ? JSON.parse(text) : null;
+  } catch {
+    json = { raw: text };
+  }
+  if (!res.ok) {
+    throw new Error(`${method} ${path} → ${res.status}: ${text}`);
+  }
+  return json;
+}
