+++ packages/cli/src/knowledge/README.md
@@ -1,0 +1,4 @@
+# Planned: `teamclu knowledge …`
+
+Reserved domain for knowledge-tree sync / import ops.
+Not implemented — the router rejects this domain until a command lands here.
