+++ packages/cli/src/mcp/README.md
@@ -1,0 +1,4 @@
+# Planned: `teamclu mcp …`
+
+Reserved domain for first-party MCP catalog / sync ops.
+Not implemented — the router rejects this domain until a command lands here.
